Partial support for message/value ports and the message context.
This use case now works:
 - Add an event input and the "print" plugin from imum.lv2 to ingen
 - Connect the event input to the input of "print"
 - Hook Ingen up to JACK and play some MIDI events
   (or get events to the print plugin from anywhere else)
 - The "print" plugin will print the received events to the console
   in the message context (i.e. the audio thread is realtime safe)


git-svn-id: http://svn.drobilla.net/lad/trunk/ingen@2281 a436a847-0d15-0410-975c-d299462d15a1

+++ src/engine/LV2BlobFeature.hpp
@@ -39,10 +39,10 @@ struct BlobFeature : public Shared::LV2Features::Feature {
 	                     LV2_Reference*        reference,
 	                     LV2_Blob_Destroy      destroy_func,
 	                     uint32_t              type,
-	                     uint32_t              size) {}
+	                     size_t                size) {}
 
 	static LV2_Blob* reference_get(LV2_Blob_Support_Data data,
-	                               LV2_Reference*    ref) { return 0; }
+	                               LV2_Reference*        ref) { return 0; }
 
 	static void reference_copy(LV2_Blob_Support_Data data,
 	                           LV2_Reference*        dst,
